backend/mcp_studio_backend/pyproject.toml
[project]
name = "mcp-studio-backend"
version = "0.1.0"
description = "Backend for MCPStudio - The Postman for Model Context Protocol"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "motor>=3.3.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "google-api-python-client>=2.117.0",
    "google-auth-oauthlib>=1.2.0",
    "websockets>=12.0",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.1.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
